PaletteCreator/pyproject.toml
[project]
name = "palettecreator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colour-science>=0.4.7",
    "matplotlib>=3.10.8",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "scipy>=1.16.3",
]
